refactor: atoms 폴더 이동
- atoms 내에서 비즈니스 로직을 사용하는 경우가 있기에 shared는 부적절
