+++ src/advanced/shared/hooks/useNotificationActions.ts
@@ -1,5 +1,5 @@
 import { useAtom } from 'jotai';
-import { addNotificationAtom, removeNotificationAtom } from '../atoms/notificationAtoms';
+import { addNotificationAtom, removeNotificationAtom } from '../../atoms/notificationAtoms';
 
 /**
  * 알림 액션을 위한 hook
